+++ R/ObjectiveTuning.R
@@ -25,6 +25,9 @@ ObjectiveTuning = R6Class("ObjectiveTuning",
     #' @field learner ([mlr3::Learner]).
     learner = NULL,
 
+     #' @field resampling ([mlr3::Resampling]).
+    resampling = NULL,
+
     #' @field measures (list of [mlr3::Measure]).
     measures = NULL,
 
@@ -46,8 +49,7 @@ ObjectiveTuning = R6Class("ObjectiveTuning",
       store_models = FALSE, allow_retrain = FALSE) {
 
       self$task = assert_task(as_task(task, clone = TRUE))
-      self$learner = assert_learner(as_learner(learner, clone = TRUE))
-      self$resampling = resampling
+      self$learner = assert_learner(as_learner(learner, clone = TRUE))      
       self$measures = assert_measures(as_measures(measures, clone = TRUE),
         task = self$task, learner = self$learner)
       self$store_benchmark_result = assert_logical(store_benchmark_result)
@@ -58,13 +60,19 @@ ObjectiveTuning = R6Class("ObjectiveTuning",
         ParamDbl$new(id = s$id, tags = ifelse(s$minimize, "minimize", "maximize"))
       }))
 
-      super$initialize(id = sprintf("%s_on_%s", self$learner$id, self$task$id), domain = self$learner$param_set,
-        codomain = codomain, check_values = check_values)
+      super$initialize(id = sprintf("%s_on_%s", self$learner$id, self$task$id), domain = self$learner$param_set, 
+        codomain = codomain, constants = ps(resampling = p_uty()), check_values = check_values)
+
+      # set resamplings in constants
+      resampling = assert_resampling(as_resampling(resampling, clone = TRUE))
+      if (!resampling$is_instantiated) resampling$instantiate(task)
+      self$resampling = resampling
+      self$constants$values$resampling = list(resampling)
     }
   ),
 
   private = list(
-    .eval_many = function(xss) {
+    .eval_many = function(xss, resampling) {
       learners = map(xss, function(x) {
         learner = self$learner$clone(deep = TRUE)
         learner$param_set$values = insert_named(learner$param_set$values, x)
@@ -114,31 +122,21 @@ ObjectiveTuning = R6Class("ObjectiveTuning",
       aggr = bmr$aggregate(self$measures)
       y = map_chr(self$measures, "id")
 
+      # add runtime
+      time = map_dbl(bmr$resample_results$resample_result, function(rr) {
+        sum(map_dbl(rr$learners, function(l) sum(l$timings)))
+      })
+      aggr[, "runtime_learners" := time]
+
       if (self$store_benchmark_result) {
         if (is.null(self$archive$benchmark_result)) {
           self$archive$benchmark_result = bmr
         } else {
           self$archive$benchmark_result$combine(bmr)
         }
-        cbind(aggr[, y, with = FALSE], uhash = bmr$uhashes)
+        cbind(aggr[, c(y, "runtime_learners"), with = FALSE], uhash = bmr$uhashes)
       } else {
-        aggr[, y, with = FALSE]
-      }
-    },
-
-    .resampling = NULL
-  ),
-
-  active = list(
-
-    #' @field resampling ([mlr3::Resampling]).
-    resampling = function(rhs) {
-      if(missing(rhs)) {
-        private$.resampling
-      } else {
-        resampling = assert_resampling(as_resampling(rhs, clone = TRUE))
-        if (!resampling$is_instantiated) resampling$instantiate(self$task)
-        private$.resampling = resampling
+        aggr[, c(y, "runtime_learners"), with = FALSE]
       }
     }
   )