+++ R/TunerDesignPoints.R
@@ -10,39 +10,44 @@
 #'
 #' @templateVar id design_points
 #' @template section_dictionary_tuners
-#' @template section_parallelization
-#' @template section_logging
 #'
 #' @inheritSection bbotk::OptimizerDesignPoints Parameters
 #' @inheritSection bbotk::OptimizerDesignPoints Progress Bars
+#' 
+#' @template section_parallelization
+#' @template section_logging
 #'
 #' @family Tuner
 #' @seealso Package \CRANpkg{mlr3hyperband} for hyperband tuning.
 #' @export
 #' @examples
-#' library(mlr3)
-#' library(paradox)
 #' library(data.table)
-#' search_space = ParamSet$new(list(
-#'   ParamDbl$new("cp", lower = 0.001, upper = 0.1)
-#' ))
-#' terminator = trm("evals", n_evals = 3)
-#' instance = TuningInstanceSingleCrit$new(
-#'   task = tsk("iris"),
-#'   learner = lrn("classif.rpart"),
+#' 
+#' # retrieve task
+#' task = tsk("pima")
+#' 
+#' # load learner and set search space
+#' learner = lrn("classif.rpart", cp = to_tune(1e-04, 1e-1, logscale = TRUE))
+#' 
+#' # hyperparameter tuning on the pima indians diabetes data set
+#' instance = tune(
+#'   method = "design_points",
+#'   task = task,
+#'   learner = learner,
 #'   resampling = rsmp("holdout"),
 #'   measure = msr("classif.ce"),
-#'   search_space = search_space,
-#'   terminator = terminator
+#'   design = data.table(cp = c(log(1e-1), log(1e-2)))
 #' )
-#' design = data.table(cp = c(0.1, 0.01))
-#' tt = tnr("design_points", design = design)
-#' # modifies the instance by reference
-#' tt$optimize(instance)
-#' # returns best configuration and best performance
+#'
+#' # best performing hyperparameter configuration
 #' instance$result
-#' # allows access of data.table of full path of all evaluations
-#' instance$archive
+#' 
+#' # all evaluated hyperparameter configuration
+#' as.data.table(instance$archive)
+#' 
+#' # fit final model on complete data set
+#' learner$param_set$values = instance$result_learner_param_vals
+#' learner$train(task)
 TunerDesignPoints = R6Class("TunerDesignPoints",
   inherit = TunerFromOptimizer,
   public = list(