+++ R/TunerNLoptr.R
@@ -17,10 +17,11 @@
 #'
 #' @templateVar id nloptr
 #' @template section_dictionary_tuners
-#' @template section_logging
 #'
 #' @inheritSection bbotk::OptimizerNLoptr Parameters
 #' @inheritSection bbotk::OptimizerNLoptr Progress Bars
+#'
+#' @template section_logging
 #'
 #' @family Tuner
 #' @seealso Package \CRANpkg{mlr3hyperband} for hyperband tuning.
@@ -30,29 +31,31 @@
 #' @export
 #' @examples
 #' \dontrun{
-#' library(mlr3)
-#' library(paradox)
-#' library(data.table)
-#' search_space = ParamSet$new(list(
-#'   ParamDbl$new("cp", lower = 0.001, upper = 0.1)
-#' ))
-#' # We use the internal termination criterion xtol_rel
-#' terminator = trm("none")
-#' instance = TuningInstanceSingleCrit$new(
-#'   task = tsk("iris"),
-#'   learner = lrn("classif.rpart"),
+#' # retrieve task
+#' task = tsk("pima")
+#' 
+#' # load learner and set search space
+#' learner = lrn("classif.rpart", cp = to_tune(1e-04, 1e-1, logscale = TRUE))
+#' 
+#' # hyperparameter tuning on the pima indians diabetes data set
+#' instance = tune(
+#'   method = "nloptr",
+#'   task = task,
+#'   learner = learner,
 #'   resampling = rsmp("holdout"),
 #'   measure = msr("classif.ce"),
-#'   search_space = search_space,
-#'   terminator = terminator
+#'   algorithm = "NLOPT_LN_BOBYQA"
 #' )
-#' tt = tnr("nloptr", algorithm = "NLOPT_LN_BOBYQA")
-#' # modifies the instance by reference
-#' tt$optimize(instance)
-#' # returns best configuration and best performance
+#'
+#' # best performing hyperparameter configuration
 #' instance$result
-#' # allows access of data.table of full path of all evaluations
-#' instance$archive
+#' 
+#' # all evaluated hyperparameter configuration
+#' as.data.table(instance$archive)
+#' 
+#' # fit final model on complete data set
+#' learner$param_set$values = instance$result_learner_param_vals
+#' learner$train(task)
 #' }
 TunerNLoptr = R6Class("TunerNLoptr",
   inherit = TunerFromOptimizer,